+++ controllers/handlers/receptor.js
@@ -1,7 +1,4 @@
-const { _M_ } = require('../modules/');
-
-
-const BrowserReceptor = (req, res, next) => {
+const browserReceptor = (req, res, next) => {                                                                           // todo: start to load template file from fs? (concurrency)
   // populating variables
   res.locals.$$MODE = 'html';
   res.locals.$$SITE = JSON.parse(process.env.$WEBSITE_CONFIGS);
@@ -14,11 +11,11 @@ const BrowserReceptor = (req, res, next) => {
     user: req.session.user,
   };
 
-  // configuring sessions
-  if (!req.session.user && req.session.cookie.expires) req.session.cookie.expires = false;
+  // housekeeping sessions
   if (res.locals.$$VIEW.flash.action[0] !== 'retry') delete req.session.returnTo;
 
-  return req.body.post && ['POST', 'PATCH'].includes(req.method) ? _M_.postNormalizer(req, res, next) : next();
+  return next();
+  // return req.body.post && ['POST', 'PATCH'].includes(req.method) ? _M_.postNormalizer(req, res, next) : next();      // tofix: move to other place
 };
 
 
@@ -31,4 +28,10 @@ const APIReceptor = (req, res, next) => {
 
 
 // exports
-module.exports = { BrowserReceptor, APIReceptor };
+module.exports = { browserReceptor, APIReceptor };
+
+Object.defineProperty(module.exports, Symbol.for('__TEST__'), {
+  value: {
+    ...module.exports,
+  },
+});